requirements.txt
tensorflow==2.15.0
pandas 
numpy
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
nltk
torch 
torchvision 
transformers 
pillow